Refactor configuration logic
Moved configration to be fetched from main backstage configuration
as opposed to getting it from annotations. Also conditional rendering
of the component as we cannot use hooks inside js functions. This was
necessary to use the useAPI hook to talk to the config API.
Will revisit how to implement this if necessary.